python/examples/advanced_agents/game_builder/grok/requirements.txt
openai
composio-openai
python-dotenv 